% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.csc_diag}
\alias{autoplot.csc_diag}
\title{Plot diagnostics}
\usage{
\method{autoplot}{csc_diag}(object, which = c("tipping", "placebo", "compare"), ...)
}
\arguments{
\item{object}{A \code{csc_diag} object.}

\item{which}{One of \code{"tipping"}, \code{"placebo"}, \code{"compare"}.}

\item{...}{Additional arguments passed to methods (currently unused).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot diagnostics
}
