% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo.R
\name{csc_placebo}
\alias{csc_placebo}
\title{Simple placebo tests}
\usage{
csc_placebo(
  df,
  y,
  d,
  x,
  type = c("random_assign", "lead_treatment", "event_lead"),
  lead_k = 1,
  seed = 123,
  did = list(id = NULL, time = NULL)
)
}
\description{
\itemize{
\item random_assign: permute treatment labels and re-fit OLS
\item lead_treatment: shift D forward by k in time-sorted data
\item event_lead: for panel with absorbing D, add k-period \emph{lead} of D in TWFE
}
}
\keyword{internal}
