% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{csc_report}
\alias{csc_report}
\title{Render a compact robustness report}
\usage{
csc_report(diag, out = "csc_report.html")
}
\arguments{
\item{diag}{csc_diag}

\item{out}{output html path}
}
\description{
Render a compact robustness report
}
