% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{csc_sensitivity}
\alias{csc_sensitivity}
\title{Sensitivity analysis: prefer sensemakr if available; fallback to simple grid}
\usage{
csc_sensitivity(mod, treat)
}
\description{
Sensitivity analysis: prefer sensemakr if available; fallback to simple grid
}
\keyword{Sensitivity}
\keyword{analysis:}
\keyword{available;}
\keyword{fallback}
\keyword{grid}
\keyword{if}
\keyword{internal}
\keyword{prefer}
\keyword{sensemakr}
\keyword{simple}
\keyword{to}
