% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{csc_sensitivity_grid}
\alias{csc_sensitivity_grid}
\title{(fallback) Sensitivity grid (tipping map) for unobserved confounding (approx)}
\usage{
csc_sensitivity_grid(
  mod,
  treat,
  r2yu = seq(0, 0.3, by = 0.01),
  r2du = seq(0, 0.3, by = 0.01)
)
}
\description{
(fallback) Sensitivity grid (tipping map) for unobserved confounding (approx)
}
\keyword{internal}
